Fix a few compiler warnings caused by type mismatches in NSLog statements.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@35037 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSData.m
@@ -1932,7 +1932,7 @@ failure:
           if (b == 0)
             {
               NSLog(@"[NSDataMalloc -initWithCoder:] unable to get %u bytes",
-                l);
+                (unsigned int)l);
               DESTROY(self);
               return nil;
             }
